requirements.txt
requests
python-dotenv
google-generativeai
reportlab
openpyxl
fastapi
uvicorn[standard]